Temp8

+++ src/Umbraco.Web/Media/EmbedProviders/OEmbedResponse.cs
@@ -31,23 +31,23 @@ namespace Umbraco.Web.Media.EmbedProviders
         public string ThumbnailUrl { get; set; }
 
         [JsonProperty("thumbnail_height")]
-        public int? ThumbnailHeight { get; set; }
+        public double? ThumbnailHeight { get; set; }
 
         [JsonProperty("thumbnail_width")]
-        public int? ThumbnailWidth { get; set; }
+        public double? ThumbnailWidth { get; set; }
 
         public string Html { get; set; }
 
         public string Url { get; set; }
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
-        public int? Width { get; set; }
+        public double? Width { get; set; }
 
         /// <summary>
         /// Gets the HTML.
         /// </summary>
-        /// <returns>The response html</returns>
+        /// <returns>The response HTML</returns>
         public string GetHtml()
         {
             if (Type == "photo")
